This fixes a bunch of warnings that are in our code, in order for the GYP build to succeed. Review URL: http://codereview.chromium.org/118346
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17865 0039d316-1c4b-4281-b951-d872f2087c98

+++ o3d/import/cross/gz_compressor.cc
@@ -40,7 +40,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include <assert.h>
 #include "import/cross/memory_buffer.h"
 #include "import/cross/memory_stream.h"
-#include "third_party/zlib/files/zutil.h"
+#include "zutil.h"
 
 const size_t kChunkSize = 16384;
 
@@ -105,7 +105,7 @@ int GzCompressor::CompressBytes(MemoryReadStream *stream,
   int result = Z_OK;
 
   // Don't try to read more than our stream has
-  int remaining = stream->GetRemainingByteCount();
+  size_t remaining = stream->GetRemainingByteCount();
   if (bytes_to_process > remaining) {
     return Z_STREAM_ERROR;
   }
